Add Teds\any(...) and all(iterable $input, ?callable $cb=null)
Moved from https://wiki.php.net/rfc/any_all_on_iterable into a pecl

- The namespace was chosen to avoid conflicting with any future additions to
  php's core itself (e.g. function signature, argument order, behavior).
- This contains functionality that was proposed for `Iterable`
  and likely to be rejected for being too small in scope
  https://wiki.php.net/rfc/cachediterable_straw_poll#namespace_choices

This has merge conflicts which were resolved in the subsequent commit(s).
This commit was used as a basis for implementing an alternate approach to
checking if all()/any() elements in an iterable satisfy a predicate.

Cherry-picked from php-src 3597 then from 6053 with modifications

Combine helpers into any() and all(), modernize for php 8

Properly free the result of the callback if it was a refcounted
type such as a string/array.

Fix the default value of the callback.

Co-Authored-By: bugreportuser
Co-Authored-By: Tyson Andre <tandre@php.net>